# daemon: paralellize disk usage computations