**- What I did**
Ensure per-object disk usage computations are performed concurrently

**- How I did it**
Wrap computations in a `errgroup.Group`.
This is that we already do on the router level https://github.com/moby/moby/blob/deda3d4933d3c0bd57f2cef672da5d28fc653706/api/server/router/system/system_routes.go#L117-L151, this just adds this functionality on the daemon level itself.  It's unfortunate that we have to maintain 2 separate groups for this, but this is just reaping the low-hanging fruit to improve performance and we can improve upon this later.

**- How to verify it**

Before:
```
root@7e6855bd30ed:/go/src/github.com/docker/docker# time docker system df
TYPE                TOTAL               ACTIVE              SIZE                RECLAIMABLE
Images              8                   0                   3.239GB             3.239GB (100%)
Containers          0                   0                   0B                  0B
Local Volumes       0                   0                   0B                  0B

real	0m0.091s
user	0m0.019s
sys	0m0.011s
root@7e6855bd30ed:/go/src/github.com/docker/docker# time docker system df
TYPE                TOTAL               ACTIVE              SIZE                RECLAIMABLE
Images              8                   0                   3.239GB             3.239GB (100%)
Containers          0                   0                   0B                  0B
Local Volumes       0                   0                   0B                  0B

real	0m0.096s
user	0m0.021s
sys	0m0.010s
root@7e6855bd30ed:/go/src/github.com/docker/docker# time docker system df
TYPE                TOTAL               ACTIVE              SIZE                RECLAIMABLE
Images              8                   0                   3.239GB             3.239GB (100%)
Containers          0                   0                   0B                  0B
Local Volumes       0                   0                   0B                  0B

real	0m0.093s
user	0m0.021s
sys	0m0.010s
```

After:
```
root@7e6855bd30ed:/go/src/github.com/docker/docker# time docker system df
TYPE                TOTAL               ACTIVE              SIZE                RECLAIMABLE
Images              8                   0                   3.239GB             3.239GB (100%)
Containers          0                   0                   0B                  0B
Local Volumes       0                   0                   0B                  0B

real	0m0.067s
user	0m0.019s
sys	0m0.010s
root@7e6855bd30ed:/go/src/github.com/docker/docker# time docker system df
TYPE                TOTAL               ACTIVE              SIZE                RECLAIMABLE
Images              8                   0                   3.239GB             3.239GB (100%)
Containers          0                   0                   0B                  0B
Local Volumes       0                   0                   0B                  0B

real	0m0.069s
user	0m0.018s
sys	0m0.012s
root@7e6855bd30ed:/go/src/github.com/docker/docker# time docker system df
TYPE                TOTAL               ACTIVE              SIZE                RECLAIMABLE
Images              8                   0                   3.239GB             3.239GB (100%)
Containers          0                   0                   0B                  0B
Local Volumes       0                   0                   0B                  0B

real	0m0.074s
user	0m0.019s
sys	0m0.011s
```